Makefile: allow for running with an arbitrary port number (#36)
The main program already has an option `--port`, but the default port
number 10000 is hard-coded in Makefile. So it's not possible to specify
an arbitrary port number. Let's define `LISTEN_PORT` in Makefile, so
you can specify the port number like that:

```
make dind LISTEN_PORT=8080
make run LISTEN_PORT=8080
```